+++ great_international/tests/test_views.py
@@ -247,6 +247,36 @@ def test_invest_home_page(document, admin_client, root_page):
     assert high_potential_ops[0]['title'] == 'Featured'
 
 
+def test_invest_sector_page(admin_client, root_page):
+    page = factories.InvestSectorPageFactory(
+        live=True, featured=True, parent=root_page
+    )
+    factories.InvestSectorPageFactory(live=True, parent=page)
+    factories.InvestSectorPageFactory(
+        live=True,
+        parent=factories.InvestSectorPageFactory()
+    )
+
+    url = reverse('api:api:pages:detail', kwargs={'pk': page.pk})
+
+    response = admin_client.get(url)
+    assert response.status_code == 200
+    assert len(response.json()['children_sectors']) == 1
+
+
+def test_invest_region_landing_page(admin_client, root_page):
+    page = factories.InvestRegionLandingPageFactory(
+        live=True, parent=root_page
+    )
+    factories.InvestSectorPageFactory(live=True, parent=page)
+
+    url = reverse('api:api:pages:detail', kwargs={'pk': page.pk})
+
+    response = admin_client.get(url)
+    assert response.status_code == 200
+    assert len(response.json()['children_sectors']) == 1
+
+
 def test_high_potential_opportunity_api(document, admin_client, root_page):
     factories.InvestHighPotentialOpportunityDetailPageFactory(
         parent=root_page,